requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pillow==11.0.0
PyPDF2==3.0.1
aiofiles==24.1.0
gunicorn==23.0.0
celery==5.4.0
redis==5.2.1
requests==2.32.5
aiofiles==24.1.0
celery==5.4.0
fastapi==0.115.6
gunicorn==23.0.0
pillow==11.0.0
PyPDF2==3.0.1
python-multipart==0.0.20
redis==5.2.1
requests==2.32.5
uvicorn[standard]==0.34.0
